services/photo-service/requirements.txt
flask
psycopg2-binary==2.9.9
werkzeug
flask-cors
